refactor(webapp): move slice state type and card type to top level in game
