+++ packages/webapp/src/game/card.d.ts
