+++ packages/webapp/src/game/store/slice/projectSlot/projectSlot.selectors.ts
@@ -1,7 +1,7 @@
 import { PlayerID } from "boardgame.io";
 import { findContribution } from "./projectSlot.utils";
 import { ProjectSlot } from "./projectSlot";
-import { JobName } from "@/game/store/slice/card";
+import { JobName } from "@/game/card";
 
 const hasWorker = (state: ProjectSlot, jobName: JobName, playerId: PlayerID): boolean => {
   const contribution = findContribution(state.contributions, jobName, playerId);
