+++ packages/webapp/src/game/store/slice/projectSlot/projectSlot.ts
@@ -2,7 +2,7 @@ import { PlayerID } from "boardgame.io";
 import { mutators } from "./projectSlot.mutators";
 import { selectors } from "./projectSlot.selectors";
 import { Contribution } from "../../../moves/type";
-import { ProjectCard } from "../card";
+import { ProjectCard } from "../../../card";
 
 export interface ProjectContribution extends Contribution {
   worker: PlayerID;
