--- packages/webapp/src/game/store/slice/type.d.ts
@@ -1,10 +1,0 @@
-export type { Deck } from './deck';
-export type { Decks } from './decks';
-export type { ActionSlot } from './actionSlot';
-export type { ActionSlots } from './actionSlots';
-export type { ProjectSlot } from './projectSlot/projectSlot';
-export type { ProjectBoard } from './projectBoard';
-export type { EventSlot } from './table';
-export type { JobSlots } from './table';
-export type { Table } from './table';
-export type { Player, Players } from './players';
